zcanic.xyz/voice_app/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0
httpx==0.24.1
openai==1.2.4
pyyaml==6.0.1
python-multipart==0.0.6
jinja2==3.1.2
loguru==0.7.0
requests==2.31.0
aiofiles==23.2.1 